requirements.txt
fastapi
uvicorn
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
jinja2
python-multipart
